Major restructure of commands. Add more error handling.
